+++ packages/azure/src/azure-openai-provider.ts
@@ -1,6 +1,10 @@
 import {
   OpenAIChatLanguageModel,
   OpenAIChatSettings,
+  OpenAIEmbeddingModel,
+  OpenAIEmbeddingSettings,
+  OpenAICompletionSettings,
+  OpenAICompletionLanguageModel,
 } from '@ai-sdk/openai/internal';
 import { loadApiKey, loadSetting } from '@ai-sdk/provider-utils';
 
@@ -13,10 +17,42 @@ export interface AzureOpenAIProvider {
   /**
 Creates an Azure OpenAI chat model for text generation.
    */
+  languageModel(
+    deploymentId: string,
+    settings?: OpenAIChatSettings,
+  ): OpenAIChatLanguageModel;
+
+  /**
+Creates an Azure OpenAI chat model for text generation.
+   */
   chat(
     deploymentId: string,
     settings?: OpenAIChatSettings,
   ): OpenAIChatLanguageModel;
+
+  /**
+Creates an Azure OpenAI model for text embeddings.
+   */
+  embedding(
+    deploymentId: string,
+    settings?: OpenAIEmbeddingSettings,
+  ): OpenAIEmbeddingModel;
+
+  /**
+Creates an Azure OpenAI model for text embeddings.
+   */
+  textEmbedding(
+    deploymentId: string,
+    settings?: OpenAIEmbeddingSettings,
+  ): OpenAIEmbeddingModel;
+
+  /**
+   * Creates an Azure OpenAI completion model for text generation.
+   */
+  completion(
+    deploymentId: string,
+    settings?: OpenAICompletionSettings,
+  ): OpenAICompletionLanguageModel;
 }
 
 export interface AzureOpenAIProviderSettings {
@@ -29,6 +65,12 @@ Name of the Azure OpenAI resource.
 API key for authenticating requests.
      */
   apiKey?: string;
+
+  /**
+Custom fetch implementation. You can use it as a middleware to intercept requests,
+or to provide a custom fetch implementation for e.g. testing.
+    */
+  fetch?: typeof fetch;
 }
 
 /**
@@ -53,21 +95,47 @@ export function createAzure(
       description: 'Azure OpenAI resource name',
     });
 
+  const url = ({ path, modelId }: { path: string; modelId: string }) =>
+    `https://${getResourceName()}.openai.azure.com/openai/deployments/${modelId}${path}?api-version=2024-05-01-preview`;
+
   const createChatModel = (
     deploymentName: string,
     settings: OpenAIChatSettings = {},
   ) =>
     new OpenAIChatLanguageModel(deploymentName, settings, {
       provider: 'azure-openai.chat',
+      url,
       headers: getHeaders,
-      url: ({ path, modelId }) =>
-        `https://${getResourceName()}.openai.azure.com/openai/deployments/${modelId}${path}?api-version=2024-05-01-preview`,
       compatibility: 'compatible',
+      fetch: options.fetch,
+    });
+
+  const createCompletionModel = (
+    modelId: string,
+    settings: OpenAICompletionSettings = {},
+  ) =>
+    new OpenAICompletionLanguageModel(modelId, settings, {
+      provider: 'azure-openai.completion',
+      url,
+      compatibility: 'compatible',
+      headers: getHeaders,
+      fetch: options.fetch,
+    });
+
+  const createEmbeddingModel = (
+    modelId: string,
+    settings: OpenAIEmbeddingSettings = {},
+  ) =>
+    new OpenAIEmbeddingModel(modelId, settings, {
+      provider: 'azure-openai.embeddings',
+      headers: getHeaders,
+      url,
+      fetch: options.fetch,
     });
 
   const provider = function (
     deploymentId: string,
-    settings?: OpenAIChatSettings,
+    settings?: OpenAIChatSettings | OpenAICompletionSettings,
   ) {
     if (new.target) {
       throw new Error(
@@ -78,7 +146,11 @@ export function createAzure(
     return createChatModel(deploymentId, settings as OpenAIChatSettings);
   };
 
+  provider.languageModel = createChatModel;
   provider.chat = createChatModel;
+  provider.completion = createCompletionModel;
+  provider.embedding = createEmbeddingModel;
+  provider.textEmbedding = createEmbeddingModel;
 
   return provider as AzureOpenAIProvider;
 }