+++ packages/ui-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/ui-utils
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [85712895]
+- Updated dependencies [85712895]
+  - @ai-sdk/provider-utils@0.0.15
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [7910ae84]
+  - @ai-sdk/provider-utils@0.0.14
+
 ## 0.0.1
 
 ### Patch Changes